Adding an unnecessary HTTP GET after HTTP DELETE.
This is an ugly patch for the bug described in #1.
I had not been able to see why after an HTTP DELETE the next request gets stalled.
I could retry the failing request, but it might actually had been processed by the server.
Therefore, I opted for doing an unnecessary but still innocuous HTTP GET.
This way, the undesired effect will happen then and the user won't even notice it as it was not expecting a result from this request.
Furthermore, the user won't suffer this effect in subsequent requests.